+++ wcm/vfs-sftp2.cpp
@@ -980,8 +980,8 @@ int FSSftp::SetFileTime ( FSPath& path, FSTime aTime, FSTime mTime, int* err, FS
 
 	LIBSSH2_SFTP_ATTRIBUTES attr;
 	attr.flags = LIBSSH2_SFTP_ATTR_ACMODTIME;
-	attr.atime = aTime;
-	attr.mtime = mTime;
+	attr.atime = (unsigned long)aTime;
+	attr.mtime = ( unsigned long )mTime;
 
 	try
 	{
